moved methods to base_tado
- changed compatibility of tadoX devices
